Fix build warnings on Fedora 26

+++ iocore/net/quic/QUICCongestionController.cc
@@ -33,7 +33,7 @@ QUICCongestionController::handle_frame(std::shared_ptr<const QUICFrame> frame)
   case QUICFrameType::ACK:
     break;
   default:
-    Debug(tag, "Unexpected frame type: %02x", frame->type());
+    Debug(tag, "Unexpected frame type: %02x", static_cast<unsigned int>(frame->type()));
     ink_assert(false);
     break;
   }
